refactor: put parser state into its own file
